% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{addmodel.mtrainer}
\alias{addmodel.mtrainer}
\title{addmodel mtrainer}
\usage{
addmodel.mtrainer(mtrainer, newmodelname)
}
\arguments{
\item{newlist}{A list of new algorithm names}
}
\value{
mtrainer object
}
\description{
addmodel mtrainer
}
