% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{auc.pcr}
\alias{auc.pcr}
\title{calculate AUC from probability of class at given rank}
\usage{
\method{auc}{pcr}(prob)
}
\arguments{
\item{prob}{list of the probability}
}
\value{
AUC
}
\description{
calculate AUC from probability of class at given rank
}
