% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranks.R
\name{auc.rank}
\alias{auc.rank}
\title{Calculate AUC using rank}
\usage{
\method{auc}{rank}(scores, y, class1 = NULL)
}
\arguments{
\item{scores}{A list of values from binary classifier}

\item{y}{A list of labels}

\item{class1}{A name of class 1}
}
\value{
the area under receiver operating curve
}
\description{
Function to calculate AUC from score and ground truth label using rank sum formula
}
\examples{
auc.rank(scores, y)
}
