% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{auprc.pcr}
\alias{auprc.pcr}
\title{calculate AUPRC from probability of class at given rank}
\usage{
auprc.pcr(prob)
}
\arguments{
\item{prob}{list of the probability}
}
\value{
AUPRC
}
\description{
calculate AUPRC from probability of class at given rank
}
