% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{build_curve_pcr}
\alias{build_curve_pcr}
\title{create ROC curve based on the rank threshold}
\usage{
build_curve_pcr(prob)
}
\arguments{
\item{prob}{list of the probability}
}
\value{
data frame with info
}
\description{
create ROC curve based on the rank threshold
}
