% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{check.pcr}
\alias{check.pcr}
\title{compare pcr with fermi-dirac distribution}
\usage{
check.pcr(pcrd)
}
\arguments{
\item{pcrd}{pcr dataframe}
}
\value{
cor, p.value, MAE, RMS, SSEV
}
\description{
compare pcr with fermi-dirac distribution
}
