% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{countXXY}
\alias{countXXY}
\title{count cases of Pxxy}
\usage{
countXXY(pcr_m, seed = 1)
}
\arguments{
\item{pcr_m}{PCR data frame}

\item{seed}{A number for random seed}
}
\value{
1 or 0
}
\description{
count cases of Pxxy
}
