% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{countXY}
\alias{countXY}
\title{count cases of Pxy}
\usage{
countXY(pcr_m, seed = 1)
}
\arguments{
\item{pcr_m}{PCR data frame}

\item{seed}{A number for random seed}
}
\value{
1 or 0
}
\description{
count cases of Pxy
}
