% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{countXYY}
\alias{countXYY}
\title{count cases of Pxyy}
\usage{
countXYY(pcr_m, seed = 1)
}
\arguments{
\item{pcr_m}{PCR data frame}

\item{seed}{A number for random seed}
}
\value{
1 or 0
}
\description{
count cases of Pxyy
}
