% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{create.auclist}
\alias{create.auclist}
\title{simple code to generate AUC list between initial and final values}
\usage{
create.auclist(initial = 0.51, final = 0.91, N = 10)
}
\arguments{
\item{initial}{the starting AUC value}

\item{final}{the ending AUC value}

\item{N}{the number of AUC list}
}
\value{
A list of AUC values
}
\description{
simple code to generate AUC list between initial and final values
}
