% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{create.labels}
\alias{create.labels}
\title{generate labels - class1 and class2}
\usage{
create.labels(N = 100, rho = 0.5)
}
\arguments{
\item{N}{size of labels}

\item{rho}{the prevalence N1/N}
}
\value{
A list of labels
}
\description{
generate labels - class1 and class2
}
