% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{create_predictions}
\alias{create_predictions}
\title{generate prediction matrix for classifer based on Gaussian score}
\usage{
create_predictions(n = 1000, m = 20, p = 0.6, auclist = NULL, y = NULL)
}
\arguments{
\item{n}{the sample size}

\item{m}{the number of AUC values}

\item{p}{the prevalence}

\item{auclist}{A list of AUC (optional)}

\item{y}{the list of labels (optional)}
}
\value{
predictions, actual_labels, actual_performance
}
\description{
generate prediction matrix for classifer based on Gaussian score
}
