% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranked.R
\name{get_fermi}
\alias{get_fermi}
\title{calculate beta, mu using normalized by N}
\usage{
get_fermi(auc, rho, N = 1, resol = 1e-04, method = "beta")
}
\arguments{
\item{auc}{A value of AUC}

\item{rho}{A value of prevalence}

\item{N}{A number of samples, when N=1, the results are normalized by N}

\item{resol}{A value for converting integral into summation}

\item{method}{'beta' for beta, mu 'lambda' for lambda1, lambda2}
}
\value{
in case of 'beta' - beta, mu, rstar
in case of 'lambda' - lambda1, lambda2, rstar
}
\description{
calculate beta, mu using normalized by N
}
