% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{new_mtrainer}
\alias{new_mtrainer}
\title{mtrainer: multiple trainer}
\usage{
new_mtrainer(x = list(), dataInfo = "temp", fitControl = NULL, update = TRUE)
}
\arguments{
\item{x}{A list of algorithm names (names(caret::getModelInfo()))}

\item{fitControl}{A list of control variables}
}
\value{
S3 object of list
}
\description{
initialize mtrainer object for multiple algorithm container
}
\examples{
t <- mtrainer(c('C5.0', 'ctree'))
}
