% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{new_pcr}
\alias{new_pcr}
\title{create dataframe of the probability of class with rank}
\usage{
new_pcr(scores, y, sample_size = 100, sample_n = 300, method = "bootstrap")
}
\arguments{
\item{scores}{list of score values}

\item{y}{list of class values}

\item{sample_size}{the number of items in one sampling selection}

\item{sample_n}{the number of realization}

\item{method}{the method for calculate probability}
}
\value{
the data class
}
\description{
create dataframe of the probability of class with rank
}
