% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr}
\alias{pcr}
\title{create S4 object of the probability of class with rank (PCR)}
\usage{
pcr(scores, y, sample_size = 100, sample_n = 300)
}
\arguments{
\item{scores}{list of score values}

\item{y}{list of class values}

\item{sample_size}{the number of items in one sampling selection}

\item{sample_n}{the number of realization}
}
\value{
S4 object
}
\description{
create S4 object of the probability of class with rank (PCR)
}
