% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr_nfold}
\alias{pcr_nfold}
\title{TODO}
\usage{
pcr_nfold(scores, y, nfold = 10)
}
\arguments{
\item{scores}{list of score values}

\item{y}{list of class values}

\item{nfold}{the number of items in one sampling selection}
}
\value{
the probability of class at given rank
}
\description{
TODO
}
