% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{pcr_sample}
\alias{pcr_sample}
\title{calculate pcr using bootstrap method}
\usage{
pcr_sample(scores, y, sample_size = 100, sample_n = 300)
}
\arguments{
\item{scores}{list of score values}

\item{y}{list of class values}

\item{sample_size}{the number of items in one sampling selection}

\item{sample_n}{the number of realization}
}
\value{
the probability of class at given rank
}
\description{
calculate pcr using bootstrap method
}
