% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{plot.mtrainer}
\alias{plot.mtrainer}
\title{plot the result of fitting models}
\usage{
\method{plot}{mtrainer}(mtrainer)
}
\arguments{
\item{mtrainer}{S3 object}
}
\value{
dot plot
}
\description{
plot the result of fitting models
}
