% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{plot.pcr}
\alias{plot.pcr}
\title{plot pcr with fermi-dirac distribution}
\usage{
\method{plot}{pcr}(pcrd, fname = "")
}
\arguments{
\item{pcrd}{pcr dataframe}
}
\value{
ggplot
}
\description{
plot pcr with fermi-dirac distribution
}
