% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{predict.mtrainer}
\alias{predict.mtrainer}
\title{predict using mtrainer models}
\usage{
\method{predict}{mtrainer}(mtrainer, newdata = NULL, class1 = NULL)
}
\arguments{
\item{mtrainer}{S3 object}

\item{newdata}{data for prediction}

\item{class1}{a name for the class 1}
}
\value{
mtrainer object
}
\description{
predict using mtrainer models
}
