% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{Pxxy.sample}
\alias{Pxxy.sample}
\title{calculate Pxxy with sampling}
\usage{
Pxxy.sample(pcr_m, iter = 5000, debug.flag = FALSE)
}
\arguments{
\item{pcr_m}{PCR data frame}

\item{iter}{A number of iteration}

\item{debug.flag}{boolean value to show plot}
}
\value{
Pxxy
}
\description{
calculate Pxxy with sampling
}
