% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{Pxy.sample}
\alias{Pxy.sample}
\title{calculate Pxy with sampling}
\usage{
Pxy.sample(pcr_m, iter = 5000, debug.flag = FALSE)
}
\arguments{
\item{pcr_m}{PCR data frame}

\item{iter}{A number of iteration}

\item{debug.flag}{boolean value to show plot}
}
\value{
Pxy (=AUC)
}
\description{
calculate Pxy with sampling
}
