% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{Pxyy.sample}
\alias{Pxyy.sample}
\title{calculate Pxyy with sampling}
\usage{
Pxyy.sample(pcr_m, iter = 5000, debug.flag = FALSE)
}
\arguments{
\item{pcr_m}{PCR data frame}

\item{iter}{A number of iteration}

\item{debug.flag}{boolean value to show plot}
}
\value{
Pxyy
}
\description{
calculate Pxyy with sampling
}
