% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{Pxyy_int}
\alias{Pxyy_int}
\title{calculate Pxyy with integral formula}
\usage{
Pxyy_int(beta, mu, rho, resol = 1e-06)
}
\arguments{
\item{beta}{A number for classifier temperature (beta)}

\item{mu}{A number for classifier chemical potential (mu)}

\item{rho}{the prevalence of dataset}

\item{resol}{the resolution of integral summation}
}
\value{
Pxyy
}
\description{
calculate Pxyy with integral formula
}
