% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{to_label}
\alias{to_label}
\title{add information about labels}
\usage{
to_label(ylist, class1 = NULL, asfactor = FALSE)
}
\arguments{
\item{ylist}{A list of labels}

\item{class1}{the label of class 1}

\item{asfactor}{A boolean value}
}
\value{
A list of labels
}
\description{
add information about labels
}
