% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{train.mtrainer}
\alias{train.mtrainer}
\title{train mtrainer models}
\usage{
\method{train}{mtrainer}(mtrainer, formula, data_list, update = FALSE, n_cores = -1)
}
\arguments{
\item{mtrainer}{S3 object}

\item{formula}{the relation to show class column name}

\item{data_list}{A list of data}

\item{update}{A boolean value}

\item{n_cores}{A number of cores for parallel process}
}
\value{
mtrainer object
}
\description{
train mtrainer models
}
