% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer.R
\name{update.mtrainer}
\alias{update.mtrainer}
\title{update mtrainer}
\usage{
\method{update}{mtrainer}(mtrainer, newlist)
}
\arguments{
\item{newlist}{A list of algorithm names}
}
\value{
mtrainer object
}
\description{
update mtrainer
}
