% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{var_auc_fermi}
\alias{var_auc_fermi}
\title{calculate variance of AUC from Fermi-Dirac distribution}
\usage{
var_auc_fermi(
  auc,
  rho,
  N = 1000,
  iter = 5000,
  method = "sampling",
  debug.flag = FALSE
)
}
\arguments{
\item{auc}{a value for AUC}

\item{rho}{the prevalence of the data set}

\item{N}{the size of the data set}

\item{iter}{the iteration number}

\item{method}{the Pxy, Pxxy, Pxyy calculation method - 'sampling', 'integral'}

\item{debug.flag}{boolean value for debug option}
}
\value{
var_auc: variance of AUC, Pxy: Pxy, Pxxy: Pxxy, Pxyy: Pxyy
}
\description{
calculate variance of AUC from Fermi-Dirac distribution
}
