% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{var_auc_pcr}
\alias{var_auc_pcr}
\title{calculate variance of AUC}
\usage{
var_auc_pcr(pcr_m, debug.flag = FALSE)
}
\arguments{
\item{pcr_m}{PCR data frame}

\item{debug.flag}{boolean value for debug option}
}
\value{
variance of AUC
}
\description{
calculate variance of AUC
}
